Follow protoactor-go's latest API
