+++ go.mod
@@ -2,9 +2,4 @@ module github.com/oklahomer/protoactor-go-future-example
 
 go 1.12
 
-require (
-	github.com/AsynkronIT/protoactor-go v0.0.0-20210125121722-bab29b9c335d
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/orcaman/concurrent-map v0.0.0-20210106121528-16402b402231 // indirect
-	github.com/serialx/hashring v0.0.0-20200727003509-22c0c7ab6b1b // indirect
-)
+require github.com/asynkron/protoactor-go v0.0.0-20220616142548-afd2d973a1d1
